+++ src/Position.java
@@ -1,48 +1,47 @@
 
 public class Position {
 
-	private int x;
-	private int y;
+	private double x;
+	private double y;
 
 	public Position() {
 		setX(0);
 		setY(0);
 	}
 
-	public Position(int tx, int ty) {
+	public Position(double tx, double ty) {
 		setX(tx);
 		setY(ty);
 	}
 
-	public void setPosition(int x, int y) {
-		double distance=Math.sqrt(Math.pow(x,2)+Math.pow(y, 2));
-		if(distance>Configuration.mapRadius-Configuration.microbeRadius) {
-			setX((int) (x*(Configuration.mapRadius-Configuration.microbeRadius)/distance));
-			setY((int) (y*(Configuration.mapRadius-Configuration.microbeRadius)/distance));
-		}
-		else {
+	public void setPosition(double x, double y) {
+		double distance = Math.sqrt(Math.pow(x, 2) + Math.pow(y, 2));
+		if (distance > Configuration.mapRadius - Configuration.microbeRadius) {
+			setX(x * (Configuration.mapRadius - Configuration.microbeRadius) / distance);
+			setY(y * (Configuration.mapRadius - Configuration.microbeRadius) / distance);
+		} else {
 			setX(x);
 			setY(y);
 		}
 	}
-	
+
 	public double distance(Position p) {
-		return Math.sqrt(Math.pow(this.x - p.x,2)+Math.pow(this.y - p.y, 2));
+		return Math.sqrt(Math.pow(this.x - p.x, 2) + Math.pow(this.y - p.y, 2));
 	}
 
-	public int getY() {
+	public double getY() {
 		return y;
 	}
 
-	public void setY(int y) {
+	public void setY(double y) {
 		this.y = y;
 	}
 
-	public int getX() {
+	public double getX() {
 		return x;
 	}
 
-	public void setX(int x) {
+	public void setX(double x) {
 		this.x = x;
 	}
 
@@ -51,12 +50,12 @@ public class Position {
 	}
 
 	public void incX() {
-		x+=10;
-		
+		x += 10;
+
 	}
 
 	public void decX() {
-		x-=10;
-		
+		x -= 10;
+
 	}
 }
